Backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Werkzeug==3.0.1
python-dotenv==1.0.0
reportlab==4.0.7
qrcode==1.0
